refactor: added generate jwt token method to Student/Sensei/Admin models. Refactored update student api to use params instead of query params.
